Fix MainWindow tabSwitcher bug
When an user would click another button-tab (e.g. Settings or Apps)
after launching libraryWidget, InkBox would reset the window entirely
and get back to the Home widget. The user would then have had to re-
click on the button-tab he wanted to pick to get the appropriate
content.

This commit fixes this issue by requiring a resetFullWindowException
boolean variable to be set to false (meaning that the user didn't click
on any button-tab before leaving libraryWidget) for the
resetWindow(true) function to occur as described in resetFullWindow(),
when libraryWidget is destroyed.